+++ Tests/KeyboardKitTests/Extensions/String+CharacterMapTests.swift
@@ -1,5 +1,5 @@
 //
-//  String+Chars.swift
+//  String+CharacterMap.swift
 //  KeyboardKit
 //
 //  Created by Daniel Saidi on 2021-12-31.
@@ -15,13 +15,5 @@ class String_CharsTests: XCTestCase {
         let string = "foo"
         let result = string.chars
         XCTAssertEqual(result, ["f", "o", "o"])
-    }
-
-    func testCharacterCollectionHasValidValues() {
-        XCTAssertEqual(String.carriageReturn, "\r")
-        XCTAssertEqual(String.newline, "\n")
-        XCTAssertEqual(String.space, " ")
-        XCTAssertEqual(String.tab, "\t")
-        XCTAssertEqual(String.zeroWidthSpace, "\u{200B}")
     }
 }