+++ RELEASE_NOTES.md
@@ -97,7 +97,8 @@ Note that the library no longer sets up English services by default, so make sur
 * `KeyboardColorReader` has been made removed. Use `Color` extensions directly instead.
 * `KeyboardSettingsUrlProvider` has been removed. Use `URL.keyboardSettings` directly instead.
 * `LocaleDirectionAnalyzer` has been removed.
-* `LocaleNameProvider` has been removed. 
+* `LocaleNameProvider` has been removed.
+* `QuotationAnalyzer` has been made removed. Use `String` extensions directly instead. 
 * `StandardCalloutActionProvider.standardProvider` has been removed.
 * `StaticKeyboardLayoutProvider` has been removed.
 * `SystemKeyboard` no longer has a `buttonContent` initializer.